refactor folder name taxonomy consistency
