+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ecomnext",
-  "version": "0.10.8",
+  "version": "0.10.9",
   "private": true,
   "scripts": {
     "dev": "next dev",
